+++ frontend/src/app/dataModel/address.model.ts
@@ -1,0 +1,6 @@
+export interface Address {
+    street: string;
+    country: string;
+    city: string;
+    postalCode: string;
+}
