+++ frontend/src/app/dataModel/index.model.ts
@@ -1,0 +1,5 @@
+export { Address } from "./address.model";
+export { User } from "./user.model";
+export { Config } from "./config.model";
+export { Sock } from "./sock.model";
+export { JWToken } from "./jwt.model";
