+++ frontend/src/app/dataModel/jwt.model.ts
@@ -1,0 +1,4 @@
+export interface JWToken {
+    expire: Date;
+    token: string;
+}
