+++ frontend/src/app/services/authService/auth.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { BackendLinkService } from '../backendservice/backend-link.service';
-import { User } from '../dataModel/user.model';
+import { User } from '../../dataModel/user.model';
 
 @Injectable({
   providedIn: 'root',
@@ -12,7 +12,6 @@ export class AuthService {
   user?: User;
 
   private error:any;
-  
 
   login(username?: string, password?: string) {
     return this.http.post<User>(this.backendLink.getLoginUrl(), {username, password}).subscribe({
@@ -31,6 +30,9 @@ export class AuthService {
   clearError() {
     this.error = undefined;
   }
+
+
+
 }
 
 // https://blog.angular-university.io/angular-jwt-authentication/
