+++ frontend/src/app/services/http-interceptors/http-error.service.ts
